+++ config/splat.us.strno0.yaml
@@ -66,7 +66,8 @@ segments:
       - [0x1D38, .data, e_room_fg]
       - [0x1DC4, .data, e_floor_trap]
       - [0x1DD4, .data, e_thornweed_corpseweed]
-      - [0x1E8C, data]
+      - [0x1E8C, .data, e_stone_skull]
+      - [0x1EB0, data]
       - [0x343C, .data, gen/rooms]
       - [0x3578, .data, gen/e_layout]
       - [0x5638, .data, stage_data]
