+++ src/st/rno0/e_stone_skull.c
@@ -1,4 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "rno0.h"
 
-INCLUDE_ASM("st/rno0/nonmatchings/e_stone_skull", EntityStoneSkull);
+#define SKULL_SPEED 5
+#define STONE_SKULL_OPACITY 0xD0
+
+#include "../e_stone_skull.h"
